Move update_deps util functions to utils module
