+++ ci_cd/utils/__init__.py
@@ -1,6 +1,18 @@
 """Utilities for CI/CD."""
 from .console_printing import Emoji
 from .file_io import update_file
-from .versions import SemanticVersion
+from .versions import (
+    SemanticVersion,
+    ignore_version,
+    parse_ignore_entries,
+    parse_ignore_rules,
+)
 
-__all__ = ["Emoji", "update_file", "SemanticVersion"]
+__all__ = [
+    "Emoji",
+    "update_file",
+    "SemanticVersion",
+    "parse_ignore_entries",
+    "parse_ignore_rules",
+    "ignore_version",
+]
